requirements.txt
discord.py>=2.0.0,<3.0.0
python-dotenv>=0.19.0,<1.0.0
aiohttp>=3.7.4,<4.0.0